add XFCE support
XFCE does not use gsettings and was always returning 'Light'.

Added a desktop detection function, and added XFCE support to theme()
and listener().

Since 'Light' was confusing for non-GTK desktops, an 'Unknown' theme
type is added so that application knows detection was unsuccessful and
can choose an appropriate response.

Added a 'watch' option to __main__ to demonstrate the listener.